backend/requirements/base.txt
django==5.0.2
django-environ==0.11.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==24.1
django-autoslug==1.9.9
django-countries==7.5.1
django-phonenumber-field==7.3.0
phonenumbers==8.13.30
drf-yasg==1.21.7
Pillow==11.1.0
argon2-cffi==23.1.0
pytz==2024.1
django-taggit==5.0.1
django-jazzmin==2.6.0
drf-spectacular==0.27.1
python-social-auth==0.3.6
social-auth-app-django==5.4.0
social-auth-core==4.5.1
redis>=5.0.0,<6.0.0
celery==5.3.4
flower==1.2.0
django-celery-beat==2.6.0
django-celery-email>=3.0.0
django-allauth==0.54.0
djangorestframework-simplejwt==5.2.2
dj-rest-auth==3.0.0
django-elasticsearch-dsl==7.4
django-elasticsearch-dsl-drf==0.22.5
whitenoise==6.6.0
gunicorn==21.2.0